The tests file was separated in various files with an appropiate name. Ref #281
